Changed from longs to ints for 32 bit types.

+++ gen-emu.h
@@ -4,10 +4,10 @@
 
 typedef signed char		sint8_t;
 typedef signed short	sint16_t;
-typedef signed long		sint32_t;
+typedef signed int		sint32_t;
 
 typedef unsigned char	uint8_t;
 typedef unsigned short	uint16_t;
-typedef unsigned long	uint32_t;
+typedef unsigned int	uint32_t;
 
 #endif /* _GEN_EMU_H_ */
